# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ProductHandler:
    properties:
      category:
        type: integer
      description:
        type: string
      imagePath1:
        type: string
      imagePath2:
        type: string
      imagePath3:
        type: string
      price:
        type: integer
      prodName:
        type: string
      quantity:
        type: integer
      size:
        type: integer
    type: object
  handlers.UpdateProduct:
    properties:
      category:
        type: integer
      description:
        type: string
      imagePath1:
        type: string
      imagePath2:
        type: string
      imagePath3:
        type: string
      price:
        type: integer
      prodName:
        type: string
      quantity:
        type: integer
      size:
        type: integer
    type: object
...
  handlers.addsoffer:
    properties:
      Amount:
        type: number
      Expire:
        type: string
      OfferName:
        type: string
      created:
        type: string
      productId:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ecom application
  title: swagger documentation for the application
  version: "1.0"
paths:
...
  /admin/SalesReport/downloadexcel:
    get:
      consumes:
      - application/json
      description: This can only be used by admin to  generate excel file of the sales
        report
      produces:
      - application/json
      responses:
        "200":
          description: Excel file generated and sent successfully
          schema:
            type: json
        "401":
          description: Failed to Generate  excel file
          schema:
            type: json
      summary: Generate Excel file  with sales report data
      tags:
      - Admin-SalesReport
  /admin/SalesReport/downloadpdf:
    get:
      consumes:
      - application/json
      description: This can only be used by admin to  generate pdf file of the sales
        report
      produces:
      - application/json
      responses:
        "200":
          description: PDF file generated and sent successfully
          schema:
            type: json
        "401":
          description: Failed to Generate  pdf file
          schema:
            type: json
      summary: Generate PDF file  with sales report data
      tags:
      - Admin-SalesReport
...
  /admin/offer:
    get:
      consumes:
      - application/json
      description: List all offer for a specific product
      produces:
      - application/json
      responses:
        "200":
          description: Listed all Offer
          schema:
            type: json
        "401":
          description: Offer not Found or Failed to List Offer
          schema:
            type: json
      summary: List offer
      tags:
      - Admin-OfferManagement
  /admin/offer/{ID}:
    post:
      consumes:
      - application/json
      description: Add an offer for a specific product by providing its ID and offer
        details
      parameters:
      - description: Product ID
        in: path
        name: ID
        required: true
        type: string
      - description: Offer details
        in: body
        name: offer
        required: true
        schema:
          $ref: '#/definitions/handlers.addsoffer'
      produces:
      - application/json
      responses:
        "200":
          description: Added new Offer
          schema:
            type: json
        "401":
          description: Offer not Found or Failed to Add New Offer
          schema:
            type: json
      summary: Add an offer for a product
      tags:
      - Admin-OfferManagement
...
  /admin/productlist:
    get:
      consumes:
      - mulipart/form-data
      description: Listing All Products
      produces:
      - application/json
      responses:
        "200":
          description: Listed All Products
          schema:
            type: json
        "401":
          description: Failed To Fetch Products
          schema:
            type: json
      summary: List Products
      tags:
      - Admin-ProductManagement
  /admin/products:
    get:
      consumes:
      - application/json
      description: Collecting the Product Details
      parameters:
      - description: Collecting Product Details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handlers.ProductHandler'
      produces:
      - application/json
      responses:
        "200":
          description: Upload Product Images
          schema:
            type: json
        "401":
          description: Failed to collect  data, please check your inputs again.
          schema:
            type: json
      summary: Add Product
      tags:
      - Admin-ProductManagement
    post:
      consumes:
      - mulipart/form-data
      description: Collecting all Images of the Product
      parameters:
      - description: Product images to upload
        in: formData
        name: images
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Product added successfully
          schema:
            type: json
        "401":
          description: Product already exist
          schema:
            type: json
      summary: Add Product Images
      tags:
      - Admin-ProductManagement
  /admin/products/{ID}:
    delete:
      consumes:
      - application/json
      description: Deleting  a specific product from the database
      parameters:
      - description: Prouct ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product Deleted successfully
          schema:
            type: json
        "401":
          description: Product cant be deleted
          schema:
            type: json
      summary: Delete Products
      tags:
      - Admin-ProductManagement
    patch:
      consumes:
      - application/json
      description: Editing all  details of a particular product
      parameters:
      - description: Prouct ID
        in: path
        name: id
        required: true
        type: integer
      - description: Edit Product Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            type: json
        "401":
          description: Failed to update Product
          schema:
            type: json
      summary: Edit Product Details
      tags:
      - Admin-ProductManagement
  /admin/usermanagement:
    get:
      consumes:
      - application/json
      description: This will list all the users
      produces:
      - application/json
      responses:
        "200":
          description: Listed all users
          schema:
            type: json
        "401":
          description: Failed to list users
          schema:
            type: json
      summary: List User
      tags:
      - Admin-UserManagement
  /admin/usermanagement/{ID}:
    patch:
      consumes:
      - application/json
      description: This will edit  a particular user's information
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            type: json
        "401":
          description: failed to update user
          schema:
            type: json
      summary: Edit User
      tags:
      - Admin-UserManagement
  /admin/usermanagement/block/{ID}:
    patch:
      consumes:
      - application/json
      description: This will block  or unblock a particular user from the system
      produces:
      - application/json
      responses:
        "200":
          description: User Status Updated
          schema:
            type: json
        "401":
          description: failed to update user status
          schema:
            type: json
      summary: Block/Unblock User
      tags:
      - Admin-UserManagement
...
swagger: "2.0"